Tweak to random generator.

+++ Tests/NULDBTestUtilities.h
@@ -11,7 +11,7 @@
 
 #define TEST_RANDOM_SEED 884432668
 
-#define Random_int_in_range(_first_, _last_) ((int)(((float)random()/(float)INT_MAX) * (_last_ - _first_) + _first_))
+#define Random_int_in_range(_first_, _last_) ((int)(((float)random()/(float)LONG_MAX) * (_last_ - _first_) + _first_))
 
 #define Random_printable_char() (Random_int_in_range(' ', '~'))
 #define Random_alpha_char() (Random_int_in_range('a', 'z'))
